+++ lib/src/widgets/keys.dart
@@ -7,3 +7,4 @@ const kStoriesFilterField = Key('stories_filter_field');
 const kDevicePreviewIcon = Key('device_preview_icon');
 const kRotateIcon = Key('rotate_icon');
 const kHideFrameIcon = Key('hide_frame_icon');
+const kStoryPinIcon = Key('story_pin_icon');