+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [next]
  - Adding support to preview the example inside a device frame 
+ - Adding pin button to story list 
 
 ## [0.1.4]
  - Adding support to Flutter 2.5